+++ test_apps/logger_app/test/sample_test.exs
@@ -1,0 +1,11 @@
+defmodule SampleTest do
+  use ExUnit.Case
+
+  test "passes" do
+    assert 1 + 1 == 2
+  end
+
+  test "also passes" do
+    assert true
+  end
+end